收藏功能

+++ partner-matching-core/src/main/java/com/partner/partnermatch/dto/FavoriteToggleRequest.java
@@ -5,5 +5,5 @@ import lombok.Data;
 @Data
 public class FavoriteToggleRequest {
     // 被收藏的用户ID
-    private Long collectUserId;
+    private Integer collectUserId;
 }
